VIITE-1359 Created own LogUtils object with time function.
