+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/util/LogUtils.scala
@@ -1,0 +1,18 @@
+package fi.liikennevirasto.digiroad2.util
+
+import org.slf4j.Logger
+
+object LogUtils {
+  val timeLoggingThresholdInMs = 1
+
+  def time[R](logger: Logger, operationName: String)(f: => R): R = {
+    val begin = System.currentTimeMillis()
+    val result = f
+    val duration = System.currentTimeMillis() - begin
+    if (duration >= timeLoggingThresholdInMs) {
+      logger.info(s"$operationName completed in $duration ms")
+    }
+    result
+  }
+
+}
